Mejoras en la presentacion

+++ SolucionMinimarket/CapaEntidad/Producto.cs
@@ -9,7 +9,7 @@ namespace CapaEntidad
         public int Id { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
-        public double Puntos { get; set; }
+        public int Puntos { get; set; }
         public int Stock { get; set; }
         public string Foto { get; set; }
         public Categoria Categoria { get; set; }
